move files and delete forest_builder_old
