+++ lukefi/metsi/app/file_io.py
@@ -5,11 +5,11 @@ from collections.abc import Callable
 from pathlib import Path
 import sqlite3
 from typing import Any, Optional
-from lukefi.metsi.data.formats.forest_centre_builder import GeoPackageBuilder, XMLBuilder
-from lukefi.metsi.data.formats.vmi10_builder import VMI10Builder
-from lukefi.metsi.data.formats.vmi11_builder import VMI11Builder
-from lukefi.metsi.data.formats.vmi12_builder import VMI12Builder
-from lukefi.metsi.data.formats.vmi13_builder import VMI13Builder
+from lukefi.metsi.data.formats.forest_centre.forest_centre_builder import GeoPackageBuilder, XMLBuilder
+from lukefi.metsi.data.formats.nfi.vmi10_builder import VMI10Builder
+from lukefi.metsi.data.formats.nfi.vmi11_builder import VMI11Builder
+from lukefi.metsi.data.formats.nfi.vmi12_builder import VMI12Builder
+from lukefi.metsi.data.formats.nfi.vmi13_builder import VMI13Builder
 
 from lukefi.metsi.data.formats.io_utils import (
     stands_to_csv_content,
@@ -19,7 +19,7 @@ from lukefi.metsi.data.formats.io_utils import (
     mela_par_file_content)
 from lukefi.metsi.app.app_io import MetsiConfiguration
 from lukefi.metsi.app.app_types import ExportableContainer
-from lukefi.metsi.data.formats.vmi9_builder import VMI9Builder
+from lukefi.metsi.data.formats.nfi.vmi9_builder import VMI9Builder
 from lukefi.metsi.domain.forestry_types import SimResults
 from lukefi.metsi.domain.forestry_types import StandList, ForestStand
 from lukefi.metsi.data.formats.declarative_conversion import Conversion
