+++ lukefi/metsi/data/formats/nfi/vmi12_builder.py
@@ -2,40 +2,40 @@ from typing import override
 from lukefi.metsi.app.utils import MetsiException
 from lukefi.metsi.data.conversion import vmi2internal
 from lukefi.metsi.data.enums.vmi import VmiIteration
-from lukefi.metsi.data.formats import util, vmi_util
+from lukefi.metsi.data.formats import util
 from lukefi.metsi.data.formats.declarative_conversion import Conversion
 from lukefi.metsi.data.formats.forest_builder_base import RowKind, VMIBuilder
-from lukefi.metsi.data.formats.vmi_const import VMI13_STAND_INDICES, VMI13_STRATUM_INDICES, VMI13_TREE_INDICES
+from lukefi.metsi.data.formats.nfi.vmi_const import VMI12_STAND_INDICES, VMI12_STRATUM_INDICES, VMI12_TREE_INDICES
+from lukefi.metsi.data.formats.nfi import vmi_util
 from lukefi.metsi.data.model import ForestStand
 from lukefi.metsi.data.vector_model import ReferenceTrees, TreeStrata
 from lukefi.metsi.domain.forestry_types import StandList
 
 
-class VMI13Builder(VMIBuilder):
+class VMI12Builder(VMIBuilder):
 
     def __init__(self, builder_flags: dict[str, bool],
                  declared_conversions: dict[str, Conversion], data_rows: list[str]) -> None:
         super().__init__(builder_flags, declared_conversions)
         for row in data_rows:
             kind = self._classify_row(row)
-            split_row = row.split()
 
             if kind == RowKind.STAND:
-                self.stand_rows.append({key: split_row[index] for key, index in VMI13_STAND_INDICES.items()})
+                self.stand_rows.append({key: row[index] for key, index in VMI12_STAND_INDICES.items()})
 
             elif kind == RowKind.STRATUM and self.builder_flags.get("strata", False):
-                stratum_row = {key: split_row[index] for key, index in VMI13_STRATUM_INDICES.items()}
+                stratum_row = {key: row[index] for key, index in VMI12_STRATUM_INDICES.items()}
                 stand_identifier = vmi_util.generate_stand_identifier(stratum_row)
                 self.stratum_rows.setdefault(stand_identifier, []).append(stratum_row)
 
             elif kind == RowKind.TREE and self.builder_flags.get("measured_trees", False):
-                tree_row = {key: split_row[index] for key, index in VMI13_TREE_INDICES.items()}
+                tree_row = {key: row[index] for key, index in VMI12_TREE_INDICES.items()}
                 stand_identifier = vmi_util.generate_stand_identifier(tree_row)
                 self.tree_rows.setdefault(stand_identifier, []).append(tree_row)
 
     @staticmethod
     def _classify_row(row: str) -> RowKind:
-        return RowKind(row[VMI13_STAND_INDICES["row_type"]])
+        return RowKind(row[VMI12_STAND_INDICES["row_type"]])
 
     @override
     def build(self) -> StandList:
@@ -48,7 +48,7 @@ class VMI13Builder(VMIBuilder):
             except Exception as e:
                 raise MetsiException(f"Parsing stand row {row} failed: {e}") from e
 
-        if self.builder_flags.get('strata', False):
+        if self.builder_flags.get("strata", False):
             for stand_identifier, stand_stratum_rows in self.stratum_rows.items():
                 strata = TreeStrata(len(stand_stratum_rows))
                 for j, stand_stratum_row in enumerate(stand_stratum_rows):
@@ -59,7 +59,7 @@ class VMI13Builder(VMIBuilder):
 
                 result[stand_identifier].tree_strata = strata
 
-        if self.builder_flags.get('measured_trees', False):
+        if self.builder_flags.get("measured_trees", False):
             for stand_identifier, stand_tree_rows in self.tree_rows.items():
                 stand_ = result.get(stand_identifier)
                 if stand_ is None:
@@ -96,20 +96,14 @@ class VMI13Builder(VMIBuilder):
         result.municipality_id = util.parse_int(vmi_util.vmi_codevalue(row["municipality"]))
         result.auxiliary_stand = row["stand_number"] != '1'
 
-        result.year = vmi_util.parse_vmi13_date(row["date"]).year
+        result.year = vmi_util.parse_vmi12_date(row["date"]).year
         result.start_year = result.year
-
-        area_ha = vmi_util.determine_vmi13_area_ha(
-            int(row["county"]),
+        area_ha = vmi_util.determine_vmi12_area_ha(
             int(row["lohkomuoto"]),
-            util.get_or_default(
-                util.parse_int(row["lohkotarkenne"]),
-                0
-            )
+            int(row["county"])
         )
-
         result.area_weight_factors = vmi_util.determine_area_factors(
-            row["osuus4m"],
+            row["osuus5m"],
             row["osuus9m"]
         )
         result.set_area(area_ha)
@@ -121,25 +115,20 @@ class VMI13Builder(VMIBuilder):
         if not lon:
             lon = util.get_or_default(util.parse_type(row["lon"], float), 0.0)
 
-        height = vmi_util.transform_vmi13_height_above_sea_level(row["height_above_sea_level"])
-        result.set_geo_location(lat, lon, height)
-
+        height = vmi_util.transform_vmi12_height_above_sea_level(row["height_above_sea_level"])
+        result.set_geo_location(lat, lon, height, "EPSG:2393")
         result.drainage_year = vmi_util.determine_drainage_year(row["ojitus_aika"], result.year)
-        result.fertilization_year = None
         result.soil_surface_preparation_year = vmi_util.determine_soil_surface_preparation_year(
             row["maanmuokkaus_aika"],
-            result.year
-        )
+            result.year)
         result.regeneration_area_cleaning_year = vmi_util.determine_clearing_of_reform_sector_year(
             row["muu_toimenpide"],
             row["muu_toimenpide_aika"],
-            result.year
-        )
+            result.year)
         result.artificial_regeneration_year = vmi_util.determine_artificial_regeneration_year(
             row["viljely"],
             row["viljely_aika"],
-            result.year
-        )
+            result.year)
         maintenance_details = vmi2internal.convert_forest_maintenance_details(
             row["hakkuu_tapa"],
             row["hakkuu_aika"],
@@ -148,8 +137,9 @@ class VMI13Builder(VMIBuilder):
         result.young_stand_tending_year = maintenance_details[0]
         result.cutting_year = maintenance_details[1]
         result.method_of_last_cutting = maintenance_details[2]
-        result.ds_main_tree_species_biological_age = vmi_util.determine_vmi13_dominant_storey_age(
-            row["vallitsevanjaksonika"]
+        result.ds_main_tree_species_biological_age = vmi_util.determine_vmi12_dominant_storey_age(
+            row["vallitsevanjakson_d13ika"],
+            row["vallitsevanjakson_ikalisays"]
         )
 
         result.peatland_type = vmi2internal.convert_peatland_forest_type(row["suotyy"])
@@ -160,17 +150,18 @@ class VMI13Builder(VMIBuilder):
         result.development_class = vmi2internal.convert_development_class(row["kehitysluokka"])
         result.main_tree_species_dominant_storey = vmi_util.determine_main_tree_species_dominant_storey(
             row["main_tree_species_dominant_storey"],
-            result.site_type_category
+            result.site_type_category,
         )
         result.basal_area = util.parse_type(row["pohjapintaala"], float)
         result.region = util.parse_int(row["county"])
 
         result.municipality_id = vmi_util.determine_municipality(
             row["municipality"],
-            row["kitukunta"]
+            row["kitukunta"],
         )
 
         if result.land_use_category and result.region is not None and result.owner_category is not None:
+
             is_ahvenanmaa = result.region == 21
             result.forest_management_category = vmi_util.determine_forest_management_category(
                 result.land_use_category,
@@ -260,7 +251,7 @@ class VMI13Builder(VMIBuilder):
         )
 
         stems_per_ha = vmi_util.determine_stems_per_ha(breast_height_diameter,
-                                                       vmi_version=VmiIteration.VMI13,
+                                                       vmi_version=VmiIteration.VMI12,
                                                        forestry_centre_id=forestry_centre_id,
                                                        ahvkeilaus=ahvkeilaus)
         origin = vmi2internal.convert_origin(row["origin"])
