+++ lukefi/metsi/data/formats/nfi/vmi9_builder.py
@@ -6,14 +6,15 @@ from lukefi.metsi.app.utils import MetsiException
 from lukefi.metsi.data.conversion import vmi2internal
 from lukefi.metsi.data.enums.internal import Origin, Storey
 from lukefi.metsi.data.enums.vmi import VmiIteration
-from lukefi.metsi.data.formats import util, vmi_util
+from lukefi.metsi.data.formats import util
 from lukefi.metsi.data.formats.declarative_conversion import Conversion
 from lukefi.metsi.data.formats.forest_builder_base import RowKind, VMIBuilder
-from lukefi.metsi.data.formats.vmi_const import (
+from lukefi.metsi.data.formats.nfi.vmi_const import (
     VMI9_STAND_COMMON,
     VMI9_STAND_INDICES_ESUOMI,
     VMI9_STAND_INDICES_PSUOMI,
     VMI9_TREE_INDICES)
+from lukefi.metsi.data.formats.nfi import vmi_util
 from lukefi.metsi.data.model import ForestStand
 from lukefi.metsi.data.vector_model import ReferenceTrees, TreeStrata
