+++ tests/data/formats/gpkg_util_test.py
@@ -3,7 +3,7 @@ import pandas as pd
 import numpy as np
 from geopandas import GeoDataFrame
 from shapely import Polygon
-from lukefi.metsi.data.formats import gpkg_util
+from lukefi.metsi.data.formats.forest_centre import gpkg_util
 from tests.data import test_util
